net: Add flower2 packet classifier based on flower and PANDA parser
This commit reimplements the flower classifier, the main consumer of
flow dissector, on top of the PANDA parser by mostly copynpastying the
flower code and modifying where the parser is used (fl_classify). The
new classifier is called "flower2". The iproute2 patch is sent
separately - but you'll notice other than replacing the user space tc
commands with "flower2", the syntax is exactly the same. This
classifier illustrates the flexibility of the PANDA parser and shows a
few simple encapsulation use cases that become convoluted and complex
because of flowdissector's intrincacies below:

- Customizing parsing behavior is impossible and requires multiple
   workarounds on client code to avoid pitfalls in special cases
   handled by flow dissector and to avoid unnecessary overhead.

- Due to its rigid nature, there's non-trivial loss of information when
   you have multiple layers of encap (eg multiple repeated ethernet
   headers, or ip headers etc).

- It is not flexible enough to map well to the semantics of hardware
   offloading of parsers i.e the software twin in the kernel and
   specific hardware semantics may have different capabilities.

This parser lets us match three levels of encapsulation and check for
properties in four stacked protocols simultaneously in a single pass.

The following call examplifies this:

```
tc filter add dev $DEV parent ffff: protocol 802.1Q prio 1 \
  flower2 vlan_ethtype ppp_ses ppp_proto ppp_ipv6 dst_ip ::1 \
  action drop
```